+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Journal des modifications
 
+## [1.2.0] - 2025-05-15
+
+### Ajouts
+- Système complet de génération de villages
+  - Pyramides aztèques avec étages et escaliers
+  - Temples japonais avec architecture shinto
+  - Châteaux normands avec donjons
+- Optimisation des performances de génération
+
+### Corrections
+- Résolution du NullPointerException dans ClientSetupHandler
+- Correction des warnings de méthodes non utilisées
+- Synchronisation client/serveur pour les structures
+
 ## [1.1.0] - 2025-05-14
 
 ### Ajouts